backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.4.0
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
